requirements.txt
# Base
ansible~=4.3.0

# Linters
yamllint~=1.26.0
flake8~=3.9.2

# Unit tests
parameterized~=0.8.1
tenacity~=8.0.1

# Molecule tests
molecule~=3.3.4
molecule-docker~=0.2.4

# Molecule verification tests
requests~=2.26.0
PyYAML~=5.4.1
pytest-testinfra~=6.4.0
